+++ spotifywebapipython/models/albumsimplified.py
@@ -84,7 +84,7 @@ class AlbumSimplified:
     def __eq__(self, other):
         try:
             return self.Name == other.Name
-        except Exception as ex:
+        except Exception:
             if (isinstance(self, AlbumSimplified )) and (isinstance(other, AlbumSimplified )):
                 return self.Name == other.Name
             return False
@@ -92,7 +92,7 @@ class AlbumSimplified:
     def __lt__(self, other):
         try:
             return self.Name < other.Name
-        except Exception as ex:
+        except Exception:
             if (isinstance(self, AlbumSimplified )) and (isinstance(other, AlbumSimplified )):
                 return self.Name < other.Name
             return False
@@ -166,6 +166,17 @@ class AlbumSimplified:
         return self._Images
 
 
+    @property
+    def ImageUrl(self) -> str:
+        """
+        Gets the first image url in the `Images` list, if images are defined;
+        otherwise, null.
+        """
+        if len(self._Images) > 0:
+            return self._Images[0].Url
+        return None
+            
+        
     @property
     def Name(self) -> str:
         """