+++ spotifywebapipython/models/artist.py
@@ -87,6 +87,17 @@ class Artist(ArtistSimplified):
 
 
     @property
+    def ImageUrl(self) -> str:
+        """
+        Gets the first image url in the `Images` list, if images are defined;
+        otherwise, null.
+        """
+        if len(self._Images) > 0:
+            return self._Images[0].Url
+        return None
+            
+        
+    @property
     def Popularity(self) -> int:
         """ 
         The popularity of the artist.