+++ spotifywebapipython/models/episodesimplified.py
@@ -94,7 +94,7 @@ class EpisodeSimplified:
     def __eq__(self, other):
         try:
             return self.Name == other.Name
-        except Exception as ex:
+        except Exception:
             if (isinstance(self, EpisodeSimplified )) and (isinstance(other, EpisodeSimplified )):
                 return self.Name == other.Name
             return False
@@ -102,7 +102,7 @@ class EpisodeSimplified:
     def __lt__(self, other):
         try:
             return self.Name < other.Name
-        except Exception as ex:
+        except Exception:
             if (isinstance(self, EpisodeSimplified )) and (isinstance(other, EpisodeSimplified )):
                 return self.Name < other.Name
             return False
@@ -193,6 +193,17 @@ class EpisodeSimplified:
         return self._Images
 
 
+    @property
+    def ImageUrl(self) -> str:
+        """
+        Gets the first image url in the `Images` list, if images are defined;
+        otherwise, null.
+        """
+        if len(self._Images) > 0:
+            return self._Images[0].Url
+        return None
+            
+        
     @property
     def IsExternallyHosted(self) -> bool:
         """