+++ spotifywebapipython/models/playlistpage.py
@@ -4,6 +4,7 @@
 from ..sautils import export
 from .pageobject import PageObject
 from .playlisttrack import PlaylistTrack
+from .track import Track
 
 @export
 class PlaylistPage(PageObject):
@@ -54,6 +55,20 @@ class PlaylistPage(PageObject):
         Array of `PlaylistTrack` objects.
         """
         return self._Items
+    
+
+    def GetTracks(self) -> list[Track]:
+        """ 
+        Gets a list of all tracks contained in the underlying `Items` list.
+        
+        This is a convenience method so one does not have to loop through the `Items`
+        array of PlaylistTrack objects to get the list of tracks.
+        """
+        result:list[Track] = []
+        item:PlaylistTrack
+        for item in self._Items:
+            result.append(item.Track)
+        return result
     
 
     def ToString(self, includeItems:bool=False) -> str: