+++ spotifywebapipython/models/playlistsimplified.py
@@ -85,7 +85,7 @@ class PlaylistSimplified:
     def __eq__(self, other):
         try:
             return self.Name == other.Name
-        except Exception as ex:
+        except Exception:
             if (isinstance(self, PlaylistSimplified )) and (isinstance(other, PlaylistSimplified )):
                 return self.Name == other.Name
             return False
@@ -93,7 +93,7 @@ class PlaylistSimplified:
     def __lt__(self, other):
         try:
             return self.Name < other.Name
-        except Exception as ex:
+        except Exception:
             if (isinstance(self, PlaylistSimplified )) and (isinstance(other, PlaylistSimplified )):
                 return self.Name < other.Name
             return False
@@ -155,6 +155,17 @@ class PlaylistSimplified:
         return self._Images
 
 
+    @property
+    def ImageUrl(self) -> str:
+        """
+        Gets the first image url in the `Images` list, if images are defined;
+        otherwise, null.
+        """
+        if len(self._Images) > 0:
+            return self._Images[0].Url
+        return None
+            
+        
     @property
     def Name(self) -> str:
         """