+++ spotifywebapipython/models/showsimplified.py
@@ -86,7 +86,7 @@ class ShowSimplified:
     def __eq__(self, other):
         try:
             return self.Name == other.Name
-        except Exception as ex:
+        except Exception:
             if (isinstance(self, ShowSimplified )) and (isinstance(other, ShowSimplified )):
                 return self.Name == other.Name
             return False
@@ -94,7 +94,7 @@ class ShowSimplified:
     def __lt__(self, other):
         try:
             return self.Name < other.Name
-        except Exception as ex:
+        except Exception:
             if (isinstance(self, ShowSimplified )) and (isinstance(other, ShowSimplified )):
                 return self.Name < other.Name
             return False
@@ -174,6 +174,17 @@ class ShowSimplified:
         return self._Images
 
 
+    @property
+    def ImageUrl(self) -> str:
+        """
+        Gets the first image url in the `Images` list, if images are defined;
+        otherwise, null.
+        """
+        if len(self._Images) > 0:
+            return self._Images[0].Url
+        return None
+            
+        
     @property
     def IsExternallyHosted(self) -> bool:
         """